requirements.txt
langgraph==0.2.28
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
chromadb==0.5.15
sentence-transformers==3.2.1
ollama==0.3.3
arxiv==2.1.3
requests==2.32.3
pypdf==5.1.0
gradio==5.6.0
python-dotenv==1.0.1
pydantic==2.9.2
tenacity==9.0.0
jupyter==1.1.1
ipykernel==6.29.5